🔥 chore(Blog.scss): remove Blog.scss file The Blog.scss file has been removed from the project. This is a chore commit as it does not add any new features or fix any bugs, but rather removes an unused file.
🔥 chore(Post.jsx): remove Post component
The Post component has been removed from the project. It was responsible for rendering a blog post from a markdown file. The functionality has been replaced with a new component that uses a different approach to render the blog posts.

🔥 chore(Post.scss): remove Post.scss file
The Post.scss file has been removed from the project. This file contained styling information for the Post component, which is no longer needed.