🎨 style(Footer.jsx, Header.jsx): update image paths The image paths in the Footer and Header components have been updated to use a relative path from the root directory of the project. This improves consistency and ensures that the images are loaded correctly regardless of the current URL.
🎨 style([slug].js): add Header and Footer components
The Header and Footer components were added to the [slug].js file. This improves the overall look and feel of the page by providing a consistent layout across all pages.

🎨 style([slug].js): add Header and Footer components to Post component
The Header and Footer components were added to the Post component to improve the overall layout and user experience of the application. Additionally, a new subheading component was added to the ReactMarkdown components to improve the readability of the markdown content.

✨ feat(Blog.module.css): add .blog__subheading style
Added a new style for the subheading of the blog post. The style includes a font size of 1.5rem, bold font weight, and a margin bottom of 10px.